requirements.txt
pybit==2.4.1
pandas==1.5.3
numpy==1.24.2
plotly==5.13.1
streamlit==1.20.0
python-dotenv==1.0.0 